requirements.txt
pyyaml>=6.0.1
portalocker>=2.8.2
python-telegram-bot>=21.0
discord.py>=2.3.2
requests>=2.31.0
